UBIFS: do not use long long with do_div
Requested by Andrew Morton: "do_div() operates on u64, not
signed long long. This will warn on several architectures".

+++ fs/ubifs/build.c
@@ -188,7 +188,7 @@ static int bud_wbuf_callback(struct ubifs_info *c, int lnum, int free, int pad)
 static int init_constants_late(struct ubifs_info *c)
 {
 	int tmp, err;
-	long long tmp64;
+	uint64_t tmp64;
 
 	c->main_bytes = c->main_lebs * c->leb_size;
